# Faster CopyFromAscii