Copy `ulong` per loop, less variable comparison, less variable changes, split multiblock into own path

Before
```
        Method |                 Type |          Mean |        Median |           RPS |
-------------- |--------------------- |-------------- |-------------- |-------------- |
 OutputHeaders |            Plaintext |   145.7818 ns |   146.7516 ns |  6,859,565.59 |
 OutputHeaders |               Common |   835.2836 ns |   842.3482 ns |  1,197,198.13 |
 OutputHeaders |              Unknown | 1,207.1645 ns | 1,198.2950 ns |    828,387.52 |
```
After
```
        Method |                 Type |          Mean |        Median |           RPS |
-------------- |--------------------- |-------------- |-------------- |-------------- |
 OutputHeaders |            Plaintext |   141.5104 ns |   141.8268 ns |  7,066,619.65 |
 OutputHeaders |               Common |   750.0226 ns |   748.9297 ns |  1,333,293.11 |
 OutputHeaders |              Unknown | 1,089.1873 ns | 1,089.6819 ns |    918,115.78 |
```

Also added test for the function...