@mharthoorn 2017-01-14
Shift8Idenitity should probably be Shift8Identity