@benaadams 2017-01-17
PRing a variant of this to coreclr https://github.com/dotnet/coreclr/pull/8969 as then will effect all serialization (json, mvc etc)