@benaadams 2017-01-20
Rebased + sp

Need to add boundary testing