@benaadams 2017-01-20
Added large boundary test range; shared parsing between single and multi block; added 64/32 bit split and alignment processing.

Probably need to rebench