@benaadams 2017-01-22
Added benchmarks

Before
```
        Method |                 Type |          Mean |     StdDev |           RPS |
-------------- |--------------------- |-------------- |----------- |-------------- |
 OutputHeaders |               Common |   745.1278 ns |  8.6515 ns |  1,342,051.58 |
 OutputHeaders | ContentLengthNumeric |    79.6235 ns |  0.5345 ns | 12,559,109.61 |
 OutputHeaders |  ContentLengthString |    74.5702 ns |  0.7695 ns | 13,410,186.63 |
 OutputHeaders |            Plaintext |   146.7861 ns |  0.9702 ns |  6,812,635.43 |
 OutputHeaders |              Unknown |   412.5315 ns |  5.5656 ns |  2,424,057.36 |
```

After
```
        Method |                 Type |          Mean |     StdDev |           RPS |
-------------- |--------------------- |-------------- |----------- |-------------- |
 OutputHeaders |               Common |   657.4066 ns |  5.1602 ns |  1,521,128.64 |
 OutputHeaders | ContentLengthNumeric |    77.6110 ns |  0.4539 ns | 12,884,770.92 |
 OutputHeaders |  ContentLengthString |    75.5184 ns |  0.6764 ns | 13,241,814.62 |
 OutputHeaders |            Plaintext |   141.2622 ns |  1.8080 ns |  7,079,035.77 |
 OutputHeaders |              Unknown |   368.7510 ns |  4.6133 ns |  2,711,857.20 |
```