@benaadams 2017-01-22
Added benchmarks

Before
```
        Method |                 Type |          Mean |        Median |           RPS |
-------------- |--------------------- |-------------- |-------------- |-------------- |
 OutputHeaders |               Common |   835.2836 ns |   842.3482 ns |  1,197,198.13 |
 OutputHeaders | ContentLengthNumeric |    79.6277 ns |    79.6925 ns | 12,558,439.91 |
 OutputHeaders |  ContentLengthString |    73.9253 ns |    74.7080 ns | 13,527,176.51 |
 OutputHeaders |            Plaintext |   145.7818 ns |   146.7516 ns |  6,859,565.59 |
 OutputHeaders |              Unknown | 1,207.1645 ns | 1,198.2950 ns |    828,387.52 |
```
After
```
        Method |                 Type |          Mean |        Median |           RPS |
-------------- |--------------------- |-------------- |-------------- |-------------- |
 OutputHeaders |               Common |   750.0226 ns |   748.9297 ns |  1,333,293.11 |
 OutputHeaders | ContentLengthNumeric |    77.5455 ns |    77.7948 ns | 12,895,660.83 |
 OutputHeaders |  ContentLengthString |    74.9223 ns |    75.2811 ns | 13,347,169.60 |
 OutputHeaders |            Plaintext |   141.5104 ns |   141.8268 ns |  7,066,619.65 |
 OutputHeaders |              Unknown | 1,089.1873 ns | 1,089.6819 ns |    918,115.78 |
```